backend/pyproject.toml
[project]
name = "capproj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "proxy-manager",
    "mitm-forwarder",
    "httpx>=0.28.1",
    "pydantic-settings>=2.11.0",
    "pydantic[email]>=2.12.4",
    "python-dotenv>=1.2.1",
]

[tool.uv.sources]
mitm-forwarder = { workspace = true }
proxy-manager = { workspace = true }

[tool.uv.workspace]
members = [
    "proxy_manager",
    "mitm_forwarder",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "module"